wirevalue

Struct Any

pub struct Any { /* private fields */ }

Represents a serialized value, wrapping the underlying serialization and deserialization details, while ensuring that we pass correctly-serialized message throughout the system.

Currently, Any passes through to bincode, but in the future we may include content-encoding information to allow for other codecs as well.

Implementations

impl Any

pub fn serialize<T: Serialize + Named>(value: &T) -> Result<Self>

Construct a new serialized value by serializing the provided T-typed value. Serialize uses the default encoding defined by the configuration key config::DEFAULT_ENCODING in the global configuration; use [serialize_with_encoding] to serialize values with a specific encoding.

pub fn serialize_as<T: Named, U: Serialize>(value: &U) -> Result<Self>

Serialize U-typed value as a T-typed value. This should be used with care (typically only in testing), as the value’s representation may be illegally coerced.

pub fn serialize_with_encoding<T: Serialize + Named>( encoding: Encoding, value: &T, ) -> Result<Self>

Serialize the value with the using the provided encoding.

pub fn serialize_with_encoding_as<T: Named, U: Serialize>( encoding: Encoding, value: &U, ) -> Result<Self>

Serialize U-typed value as a T-typed value. This should be used with care (typically only in testing), as the value’s representation may be illegally coerced.

pub fn new_broken() -> Self

Create a new broken Any value. A broken value has unknown type and no valid data. Attempting to deserialize a broken value will fail.

pub fn is_broken(&self) -> bool

Returns true if this Any is broken (unknown type, no value).

pub fn deserialized<T: DeserializeOwned + Named>(&self) -> Result<T>

Deserialize a value to the provided type T.

pub fn deserialized_unchecked<T: DeserializeOwned>(&self) -> Result<T>

Deserialize a value to the provided type T, without checking for type conformance. This should be used carefully, only when you know that the dynamic type check is not needed.

pub fn transcode_to_json(self) -> Result<Self, Self>

Transcode the serialized value to JSON. This operation will succeed if the type hash is embedded in the value, and the corresponding type is available in this binary.

pub fn dump(&self) -> Result<Value>

Dump the Any message into a JSON value. This will succeed if: 1) the typehash is embedded in the serialized value; 2) the named type is linked into the binary.

pub fn encoding(&self) -> Encoding

The encoding used by this serialized value.

pub fn typehash(&self) -> u64

The typehash of the serialized value.

pub fn typename(&self) -> Option<&'static str>

The typename of the serialized value, if available.

pub fn prefix<T: DeserializeOwned>(&self) -> Result<T>

Deserialize a prefix of the value. This is currently only supported for bincode-serialized values.

pub fn emplace_prefix<T: Serialize + DeserializeOwned>( &mut self, prefix: T, ) -> Result<()>

Emplace a new prefix to this value. This is currently only supported for bincode-serialized values.

pub fn len(&self) -> usize

The length of the underlying serialized message

pub fn is_empty(&self) -> bool

Is the message empty. This should always return false.

pub fn crc(&self) -> u32

Returns the 32bit crc of the serialized data

pub fn is<M: Named>(&self) -> bool

Returns whether this value contains a serialized M-typed value. Returns None when type information is unavailable.